accton-asgvolt64.conf: do not autoload onlp kmods
ONLP has a very odd way of discovering what the platform is and
installing the correct kernel mods and loading them at
first boot.

tbh, I'm not thrilled with how they do it as it makes a RO
rootfs impossible. We'll eventually need to fix this, but
for now, it works as is written on the box.

I've also added the one onlp out of tree kmod needed optoe

Signed-off-by: Eilís Ní Fhlannagáin <pidge@toganlabs.com>