PORTING.md: Revise porting guide
A number of changes were made, primarily applying the recommendations
of the reviewers of the initial commit, to mirror the style of other
mion documentation, and to provide more examples. markdownlint was
also ran on the file, and a small number of syntax and style issues
were fixed.

It is possible that further changes will be needed including removing
content not seen as necessary but added in this commit, and fact
checking of technical statements.

This applies to issue #6

signed-off-by: Katrina Prosise <igorina@toganlabs.com>